Adds Grunt and related tools to the Pantheon deploy process (more)
This commit does quite a bit, and frankly I'm not sure it is properly
scoped. Right now it adds two bash scripts to both the "build" and
"build and deploy" steps on CircleCI. The first script installs the
needed tooling for our build process: Node, Grunt, and SASS. The second
script then recursively checks for themes that need Grunt, and runs
whatever steps are called for in those themes.

Questions at this point:
1. Should we be using a different Docker image that has Node/Grunt/etc
   already? Or is it okay to install them this way?
2. Is it really necessary to explicitly set our scripts to be
   executable? I got error code 126 without these steps, but none of the
   other scripts in this repo need that step.
3. Should there be a common install step that the various build scripts
   rely on? There's a fair amount of repetition currently between steps.